The asRDMol() method of MolpherMol now returns the actual pointer to the underlying rdkit molecule instance (used to be a copy before, which was not very efficient)

+++ src/core/chem/ChemicalAuxiliary.cpp
@@ -28,9 +28,8 @@ AtomIdx GetRandomAtom(const std::vector<MolpherAtom> &atoms, RDKit::Atom &atom)
 {
     int idx = SynchRand::GetRandomNumber(atoms.size() - 1);
 
-    atom.setAtomicNum(atoms[idx].getAtomicNum());
+    atom.setAtomicNum((int) atoms[idx].getAtomicNum());
     atom.setFormalCharge(atoms[idx].getFormalCharge());
-//    atom.setMass(atoms[idx].mass); // removed in new version of rdkit (mass determined from atomic number)
 
     return (AtomIdx) idx;
 }
